td: add package
Add telegram library for building Telegram clients.
This can be coming handy if you need a Telegram bot
running on your OpenWrt device.
Building the libraries requires building a host tool first.
This is because it uses it's own mtproto which source codes
are autogenerated from *.mt files on the host before
cross compilation can happen.

Signed-off-by: Toni Uhlig <matzeton@googlemail.com>